refactor: move query that is used in multiple features
