+++ src/features/offer/pages/Offer/Offer.native.test.tsx
@@ -6,13 +6,13 @@ import {
 } from 'api/gen'
 import { bookingsSnap } from 'features/bookings/fixtures'
 import { offerResponseSnap } from 'features/offer/fixtures/offerResponse'
-import * as useArtistResultsAPI from 'features/offer/queries/useArtistResultsQuery'
 import * as useSimilarOffersAPI from 'features/offer/queries/useSimilarOffersQuery'
 import { renderOfferPage } from 'features/offer/tests/renderOfferPageTestUtil'
 import { mockedAlgoliaOffersWithSameArtistResponse } from 'libs/algolia/fixtures/algoliaFixtures'
 import { setFeatureFlags } from 'libs/firebase/firestore/featureFlags/tests/setFeatureFlags'
 import { RemoteStoreFeatureFlags } from 'libs/firebase/firestore/types'
 import { PLACEHOLDER_DATA } from 'libs/subcategories/placeholderData'
+import * as useArtistResultsAPI from 'queries/offer/useArtistResultsQuery'
 import { mockServer } from 'tests/mswServer'
 import { screen, userEvent, waitFor } from 'tests/utils'
 import * as useModal from 'ui/components/modals/useModal'
