+++ src/queries/offer/fetchOffersByArtist.ts
